move TabBar to own file
